requirements.txt
djangorestframework==3.8.2
pillow==5.2.0
requests==2.20.1
django==1.11.15
psycopg2==2.7.4
ruamel.yaml==0.15.50
django-cors-headers==2.4.0
django_filter==2.0.0
Celery==4.2.* # TODO Will be used for threads
redis==2.10.6
django-debug-toolbar==1.9.1
scipy==1.1.0
pandas==0.23.4
Cython==0.27.3
setuptools==39.1.0
scikit-learn==0.19.1
tqdm==4.25.0
auto-sklearn==0.4.2
tpot==0.9.5
billiard==3.5.0.4 # Subprocesses for celery
ptvsd==3.0.0  # Debugging
autokeras==0.3.5
pyarrow==0.11.1
opencv-contrib-python==4.0.0.21
librosa==0.6.2
pyamg==4.0.0
imageio==2.5.0